requirements.txt
langgraph 
langchain[openai]
langchain-community 
langchain-text-splitters
langchain-voyageai
langchain-pinecone
langgraph-checkpoint-postgres
pypdf
boto3
unstructured

fastapi[standard]
psycopg[binary,pool]

pytest
SQLAlchemy
alembic
python-dotenv

